Dockerfile: Finish installing dependencies
I may have to prune down the list a bit later on, but for now this
takes care of most of the list.

It is probably better to install the specific per-project dependencies
down inside of each layer, which I will do later on once the rest of
the steps are fleshed out.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>