Dockerfile: Split dev/rt dependencies
We don't need to have header files and .pc files in the base image, so
avoid adding those by instead just installing the basics in the base
image and install the development crap in the build-deps image.

To build a nice <300MiB image that uses the built artifacts, we can
build off the build-deps target, but then just COPY the files into
something that keys off the base image. (Hopefully.)

If we need to do runtime compilation, that all goes out the window.  I
hope it doesn't come to that.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>