Dockerfile: Also install ospd-openvas, run that instead
`ospd-openvas` is apparently the new wrapper around `openvas`,
replacing its old daemon functionality in a more "modern"
language (Python).

This commit adds `python3` and `python3-pip` pre-emptively to _all_
packages (yikes, but at this point disk usage doesn't really matter to
me as much) and then adds the `ospd==2.0.0` and `ospd-openvas==1.0.0`
packages to the openvas container, before modifying the entrypoint to
point to the new binary (/usr/local/bin/ospd-openvas) with debugging
turned on and a log file chosen.

Note that `ospd-openvas` will hard-crash if all of the following are
true (which is the case by default in Docker)

- Foreground mode is not specified (can be turned on with a flag)
- A log file is not specified (as demonstrated in this commit)
- /dev/log does not exist or cannot be Syslogged to.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>